Add privilege checks galore

+++ htdocs/js/pages/admin/AlertSetup.class.js
@@ -11,6 +11,7 @@ Page.AlertSetup = class AlertSetup extends Page.PageUtils {
 	onActivate(args) {
 		// page activation
 		if (!this.requireLogin(args)) return true;
+		if (!this.requireAnyPrivilege('create_alerts', 'edit_alerts', 'delete_alerts')) return true;
 		
 		if (!args) args = {};
 		if (!args.sub) args.sub = this.default_sub;
@@ -66,15 +67,15 @@ Page.AlertSetup = class AlertSetup extends Page.PageUtils {
 		
 		html += this.getBasicGrid( grid_opts, function(item, idx) {
 			var actions = [];
-			actions.push( '<span class="link" onClick="$P().edit_alert('+idx+')"><b>Edit</b></span>' );
-			actions.push( '<span class="link danger" onClick="$P().delete_alert('+idx+')"><b>Delete</b></span>' );
+			if (app.hasPrivilege('edit_alerts')) actions.push( '<span class="link" onClick="$P().edit_alert('+idx+')"><b>Edit</b></span>' );
+			if (app.hasPrivilege('delete_alerts')) actions.push( '<span class="link danger" onClick="$P().delete_alert('+idx+')"><b>Delete</b></span>' );
 			
 			var tds = [
 				'<div class="td_drag_handle" style="cursor:default">' + self.getFormCheckbox({
 					checked: item.enabled,
 					onChange: '$P().toggle_alert_enabled(this,' + idx + ')'
 				}) + '</div>',
-				'<b>' + self.getNiceAlert(item, true) + '</b>',
+				'<b>' + self.getNiceAlert(item, app.hasPrivilege('edit_alerts')) + '</b>',
 				'<span class="mono">' + item.id + '</span>',
 				self.getNiceGroupList(item.groups, '', 3),
 				self.getNiceUser(item.username, app.isAdmin()),
@@ -89,9 +90,9 @@ Page.AlertSetup = class AlertSetup extends Page.PageUtils {
 		html += '</div>'; // box_content
 		
 		html += '<div class="box_buttons">';
-			html += '<div class="button" onClick="$P().doFileImportPrompt()"><i class="mdi mdi-cloud-upload-outline">&nbsp;</i>Import File...</div>';
+			if (app.hasAnyPrivilege('create_alerts', 'edit_alerts')) html += '<div class="button" onClick="$P().doFileImportPrompt()"><i class="mdi mdi-cloud-upload-outline">&nbsp;</i>Import File...</div>';
 			html += '<div class="button secondary" onClick="$P().go_history()"><i class="mdi mdi-history">&nbsp;</i>Revision History...</div>';
-			html += '<div class="button default" onClick="$P().edit_alert(-1)"><i class="mdi mdi-plus-circle-outline">&nbsp;</i>New Alert...</div>';
+			if (app.hasPrivilege('create_alerts')) html += '<div class="button default" onClick="$P().edit_alert(-1)"><i class="mdi mdi-plus-circle-outline">&nbsp;</i>New Alert...</div>';
 		html += '</div>'; // box_buttons
 		
 		html += '</div>'; // box
